+++ govstack_api/building_blocks/bb_digital_registries/views/__init__.py
@@ -1,0 +1,8 @@
+from .error_handler import handle_mutation_exceptions
+from .personal_data_view import PersonalDataAPI
+from .multiple_record_view import MultipleRecordAPI
+from .update_or_create_record_view import UpdateOrCreateRecordView
+from .record_exist_view import CheckRecordPresenceView
+from .search_record_view import SearchRecordView
+from .single_record_view import SingleRecordAPI
+
