+++ govstack_api/building_blocks/bb_digital_registries/views/personal_data_view.py
@@ -1,0 +1,7 @@
+from django.http import HttpResponse
+from rest_framework.views import APIView
+
+
+class PersonalDataAPI(APIView):
+    def get(self, request):
+        return HttpResponse(status=204)  # 204 No Content
